Don't feed embedded records to the Primary Array
Reference: https://github.com/emberjs/data/blob/48c02654e8a524b390d1a28975416ee73f912d9e/packages/ember-data/lib/serializers/rest_serializer.js#L401-L408

extractArray, when responding to findQuery or findMany must ensure that
sideloaded records aren't returned in the primary array. Otherwise, the
sideloaded records will be interpreted as results of the query.

This ends up being a serious problem with reflexive relationships.

For example, assume the following API, which models database replication

    GET /databases/10
    {
      "id": 10,
      "handle": "redis-master",
      "type": "redis",
      "_links": {
        "self": { "href": "/databases/10" },
        "dependents": { "href": "/databases/10/dependents" }
      }
    }

    GET /databases/11
    {
      "id": 11,
      "handle": "redis-replica",
      "type": "redis",
      "_embedded": {
        "initialize_from": {
          << RECORD 10 EMBEDDED HERE >>
        }
      },
      "_links": {
        "self": { "href": "/databases/11" }
      }
    }

    GET /databases/10/dependents
    {
      "_embedded": {
        "dependents": [
          << RECORD 11 EMBEDDED HERE >>
        ]
      }
    }

Wht the current behavior, when you query `/databases/10/dependents`,
then database 11 is embedded, and so is database 10. Since those
are all database records, they will all be returned in the "databases"
field of the result.

Since that's the primary array, those records will be interpreted as
dependents of database 10, and database 10 ends up being a dependent of
itself.

The Ember code suggests that "databases" should be prefixed with "_" in
this case, to ensure that sideloads are properly interpreted as
sideloads, and not as members of the primary array.